buy airtime added

+++ lib/src/features/agent/home/domain/models/agt_profile_model.dart
@@ -58,7 +58,7 @@ class DataData {
   String name;
   String email;
   String walletId;
-  double walletBalance;
+  int walletBalance;
   bool isVerified;
   String phoneNumber;
   String deviceId;
